+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## HEAD
 
+* Ensure password is only updated via #change or #reset.
 * Move views into gem's app/views/ directory.